+++ src/compose.ts
@@ -1,7 +1,12 @@
 import type { AnyElysia } from '.'
 
 import { Value } from '@sinclair/typebox/value'
-import { TypeBoxError, type TAnySchema, type TSchema } from '@sinclair/typebox'
+import {
+	Kind,
+	TypeBoxError,
+	type TAnySchema,
+	type TSchema
+} from '@sinclair/typebox'
 
 import { parseQuery, parseQueryFromURL } from './fast-querystring'
 
@@ -344,6 +349,69 @@ export const hasProperty = (
 
 const TransformSymbol = Symbol.for('TypeBox.Transform')
 
+export const hasRef = (schema: TAnySchema): boolean => {
+	if (!schema) return false
+
+	if (schema.type === 'object' && schema.properties) {
+		const properties = schema.properties as Record<string, TAnySchema>
+		for (const key of Object.keys(properties)) {
+			const property = properties[key]
+
+			if (property.type === 'object') {
+				if (hasRef(property)) return true
+			} else {
+				if (property.anyOf) {
+					for (let i = 0; i < property.anyOf.length; i++)
+						if (hasRef(property.anyOf[i])) return true
+				}
+
+				if (property.oneOf) {
+					for (let i = 0; i < property.oneOf.length; i++)
+						if (hasRef(property.oneOf[i])) return true
+				}
+
+				if (property.allOf) {
+					for (let i = 0; i < property.allOf.length; i++)
+						if (hasRef(property.allOf[i])) return true
+				}
+			}
+
+			if (property[Kind] === 'Ref' && '$ref' in property) return true
+
+			if (property.type === 'array' && property.items) {
+				const item = property.items
+
+				if (item.type === 'object') {
+					if (hasRef(item)) return true
+				} else {
+					if (item.anyOf) {
+						for (let i = 0; i < item.anyOf.length; i++)
+							if (hasRef(item.anyOf[i])) return true
+					}
+
+					if (item.oneOf) {
+						for (let i = 0; i < item.oneOf.length; i++)
+							if (hasRef(item.oneOf[i])) return true
+					}
+
+					if (item.allOf) {
+						for (let i = 0; i < item.allOf.length; i++)
+							if (hasRef(item.allOf[i])) return true
+					}
+				}
+
+				if (item[Kind] === 'Ref' && '$ref' in item) return true
+			}
+		}
+
+		return false
+	}
+
+	if (schema.type === 'array' && schema.items) return hasRef(schema.items)
+
+	return schema[Kind] === 'Ref' && '$ref' in schema
+}
+
 export const hasTransform = (schema: TAnySchema): boolean => {
 	if (!schema) return false
 
@@ -357,17 +425,56 @@ export const hasTransform = (schema: TAnySchema): boolean => {
 
 			if (property.type === 'object') {
 				if (hasTransform(property)) return true
-			} else if (property.anyOf) {
-				for (let i = 0; i < property.anyOf.length; i++)
-					if (hasTransform(property.anyOf[i])) return true
+			} else {
+				if (property.anyOf) {
+					for (let i = 0; i < property.anyOf.length; i++)
+						if (hasTransform(property.anyOf[i])) return true
+				}
+
+				if (property.allOf) {
+					for (let i = 0; i < property.allOf.length; i++)
+						if (hasTransform(property.allOf[i])) return true
+				}
+
+				if (property.oneOf) {
+					for (let i = 0; i < property.oneOf.length; i++)
+						if (hasTransform(property.oneOf[i])) return true
+				}
 			}
 
-			const hasTransformSymbol = TransformSymbol in property
-			if (hasTransformSymbol) return true
+			if (TransformSymbol in property) return true
+
+			if (property.type === 'array' && property.items) {
+				const item = property.items
+
+				if (item.type === 'object') {
+					if (hasTransform(item)) return true
+				} else {
+					if (item.anyOf) {
+						for (let i = 0; i < item.anyOf.length; i++)
+							if (hasTransform(item.anyOf[i])) return true
+					}
+
+					if (item.oneOf) {
+						for (let i = 0; i < item.oneOf.length; i++)
+							if (hasTransform(item.oneOf[i])) return true
+					}
+
+					if (item.allOf) {
+						for (let i = 0; i < item.allOf.length; i++)
+							if (hasTransform(item.allOf[i])) return true
+					}
+				}
+
+				if (TransformSymbol in item) return true
+			}
 		}
 
 		return false
 	}
+
+	if (schema.type === 'array' && schema.items)
+		return hasTransform(schema.items)
 
 	return (
 		TransformSymbol in schema ||